[CDNG-5668]: Access control for secrets (#22280)
* [CDNG-5668]: Access control for secret APIs

+++ 120-ng-manager/src/main/java/io/harness/ng/core/remote/NGSecretResourceV2.java
@@ -1,12 +1,25 @@
 package io.harness.ng.core.remote;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
+import static io.harness.exception.WingsException.USER;
+import static io.harness.secrets.SecretPermissions.SECRET_ACCESS_PERMISSION;
+import static io.harness.secrets.SecretPermissions.SECRET_DELETE_PERMISSION;
+import static io.harness.secrets.SecretPermissions.SECRET_EDIT_PERMISSION;
+import static io.harness.secrets.SecretPermissions.SECRET_RESOURCE_TYPE;
+import static io.harness.secrets.SecretPermissions.SECRET_VIEW_PERMISSION;
 
 import io.harness.NGCommonEntityConstants;
 import io.harness.NGResourceFilterConstants;
+import io.harness.accesscontrol.clients.AccessControlClient;
+import io.harness.accesscontrol.clients.Resource;
+import io.harness.accesscontrol.clients.ResourceScope;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.beans.DecryptableEntity;
 import io.harness.data.validator.EntityIdentifier;
+import io.harness.encryption.SecretRefData;
+import io.harness.exception.InvalidRequestException;
 import io.harness.ng.beans.PageResponse;
+import io.harness.ng.core.NGAccess;
 import io.harness.ng.core.NGAccessWithEncryptionConsumer;
 import io.harness.ng.core.api.SecretCrudService;
 import io.harness.ng.core.dto.ErrorDTO;
@@ -27,7 +40,10 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import java.io.InputStream;
+import java.lang.reflect.Field;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
@@ -46,6 +62,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.glassfish.jersey.media.multipart.FormDataParam;
 import org.hibernate.validator.constraints.NotEmpty;
 
@@ -61,11 +78,13 @@ import org.hibernate.validator.constraints.NotEmpty;
     })
 @AllArgsConstructor(onConstructor = @__({ @Inject }))
 @NextGenManagerAuth
+@Slf4j
 public class NGSecretResourceV2 {
   private static final String INCLUDE_SECRETS_FROM_EVERY_SUB_SCOPE = "includeSecretsFromEverySubScope";
   private final SecretCrudService ngSecretService;
   private final Validator validator;
   private final SecretManagerClientService secretManagerClientService;
+  private final AccessControlClient accessControlClient;
 
   @GET
   @Path("/validateUniqueIdentifier/{identifier}")
@@ -84,7 +103,17 @@ public class NGSecretResourceV2 {
   @ApiOperation(value = "Create a secret", nickname = "postSecret")
   public ResponseDTO<SecretResponseWrapper> create(
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) @NotNull String accountIdentifier,
+      @QueryParam(NGCommonEntityConstants.ORG_KEY) String orgIdentifier,
+      @QueryParam(NGCommonEntityConstants.PROJECT_KEY) String projectIdentifier,
       @Valid @NotNull SecretRequestWrapper dto) {
+    if (!Objects.equals(orgIdentifier, dto.getSecret().getOrgIdentifier())
+        || !Objects.equals(projectIdentifier, dto.getSecret().getProjectIdentifier())) {
+      throw new InvalidRequestException("Invalid request, scope in payload and params do not match.", USER);
+    }
+
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier), Resource.NONE, SECRET_EDIT_PERMISSION);
+
     return ResponseDTO.newResponse(ngSecretService.create(accountIdentifier, dto.getSecret()));
   }
 
@@ -97,6 +126,8 @@ public class NGSecretResourceV2 {
       @QueryParam(NGCommonEntityConstants.ORG_KEY) String orgIdentifier,
       @QueryParam(NGCommonEntityConstants.PROJECT_KEY) String projectIdentifier,
       @QueryParam(NGCommonEntityConstants.IDENTIFIER_KEY) String identifier, @Valid SecretValidationMetaData metadata) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier),
+        Resource.of(SECRET_RESOURCE_TYPE, identifier), SECRET_ACCESS_PERMISSION);
     return ResponseDTO.newResponse(
         ngSecretService.validateSecret(accountIdentifier, orgIdentifier, projectIdentifier, identifier, metadata));
   }
@@ -107,7 +138,15 @@ public class NGSecretResourceV2 {
   @ApiOperation(value = "Create a secret via yaml", nickname = "postSecretViaYaml")
   public ResponseDTO<SecretResponseWrapper> createViaYaml(
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) @NotNull String accountIdentifier,
+      @QueryParam(NGCommonEntityConstants.ORG_KEY) @NotNull String orgIdentifier,
+      @QueryParam(NGCommonEntityConstants.PROJECT_KEY) @NotNull String projectIdentifier,
       @Valid SecretRequestWrapper dto) {
+    if (!Objects.equals(orgIdentifier, dto.getSecret().getOrgIdentifier())
+        || !Objects.equals(projectIdentifier, dto.getSecret().getProjectIdentifier())) {
+      throw new InvalidRequestException("Invalid request, scope in payload and params do not match.", USER);
+    }
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier), Resource.NONE, SECRET_EDIT_PERMISSION);
     return ResponseDTO.newResponse(ngSecretService.createViaYaml(accountIdentifier, dto.getSecret()));
   }
 
@@ -124,6 +163,9 @@ public class NGSecretResourceV2 {
       @QueryParam(INCLUDE_SECRETS_FROM_EVERY_SUB_SCOPE) @DefaultValue("false") boolean includeSecretsFromEverySubScope,
       @QueryParam(NGResourceFilterConstants.PAGE_KEY) @DefaultValue("0") int page,
       @QueryParam(NGResourceFilterConstants.SIZE_KEY) @DefaultValue("100") int size) {
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier), Resource.NONE, SECRET_VIEW_PERMISSION);
+
     if (secretType != null) {
       secretTypes.add(secretType);
     }
@@ -139,6 +181,8 @@ public class NGSecretResourceV2 {
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) @NotNull String accountIdentifier,
       @QueryParam(NGCommonEntityConstants.ORG_KEY) String orgIdentifier,
       @QueryParam(NGCommonEntityConstants.PROJECT_KEY) String projectIdentifier) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier),
+        Resource.of(SECRET_RESOURCE_TYPE, identifier), SECRET_VIEW_PERMISSION);
     return ResponseDTO.newResponse(
         ngSecretService.get(accountIdentifier, orgIdentifier, projectIdentifier, identifier).orElse(null));
   }
@@ -150,6 +194,8 @@ public class NGSecretResourceV2 {
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) @NotNull String accountIdentifier,
       @QueryParam(NGCommonEntityConstants.ORG_KEY) String orgIdentifier,
       @QueryParam(NGCommonEntityConstants.PROJECT_KEY) String projectIdentifier) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier),
+        Resource.of(SECRET_RESOURCE_TYPE, identifier), SECRET_DELETE_PERMISSION);
     return ResponseDTO.newResponse(
         ngSecretService.delete(accountIdentifier, orgIdentifier, projectIdentifier, identifier));
   }
@@ -163,6 +209,8 @@ public class NGSecretResourceV2 {
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) @NotNull String accountIdentifier,
       @QueryParam(NGCommonEntityConstants.ORG_KEY) String orgIdentifier,
       @QueryParam(NGCommonEntityConstants.PROJECT_KEY) String projectIdentifier, @Valid SecretRequestWrapper dto) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier),
+        Resource.of(SECRET_RESOURCE_TYPE, identifier), SECRET_EDIT_PERMISSION);
     return ResponseDTO.newResponse(
         ngSecretService.update(accountIdentifier, orgIdentifier, projectIdentifier, identifier, dto.getSecret()));
   }
@@ -176,6 +224,9 @@ public class NGSecretResourceV2 {
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) @NotNull String accountIdentifier,
       @QueryParam(NGCommonEntityConstants.ORG_KEY) String orgIdentifier,
       @QueryParam(NGCommonEntityConstants.PROJECT_KEY) String projectIdentifier, @Valid SecretRequestWrapper dto) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier),
+        Resource.of(SECRET_RESOURCE_TYPE, identifier), SECRET_EDIT_PERMISSION);
+
     return ResponseDTO.newResponse(ngSecretService.updateViaYaml(
         accountIdentifier, orgIdentifier, projectIdentifier, identifier, dto.getSecret()));
   }
@@ -198,8 +249,10 @@ public class NGSecretResourceV2 {
       @PathParam(NGCommonEntityConstants.IDENTIFIER_KEY) String identifier,
       @FormDataParam("file") InputStream uploadedInputStream, @FormDataParam("spec") String spec) {
     SecretRequestWrapper dto = JsonUtils.asObject(spec, SecretRequestWrapper.class);
-
     validateRequestPayload(dto);
+
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier),
+        Resource.of(SECRET_RESOURCE_TYPE, identifier), SECRET_EDIT_PERMISSION);
 
     return ResponseDTO.newResponse(ngSecretService.updateFile(
         accountIdentifier, orgIdentifier, projectIdentifier, identifier, dto.getSecret(), uploadedInputStream));
@@ -211,10 +264,19 @@ public class NGSecretResourceV2 {
   @Consumes(MediaType.MULTIPART_FORM_DATA)
   public ResponseDTO<SecretResponseWrapper> createSecretFile(
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) @NotNull String accountIdentifier,
+      @QueryParam(NGCommonEntityConstants.ORG_KEY) String orgIdentifier,
+      @QueryParam(NGCommonEntityConstants.PROJECT_KEY) String projectIdentifier,
       @NotNull @FormDataParam("file") InputStream uploadedInputStream, @FormDataParam("spec") String spec) {
     SecretRequestWrapper dto = JsonUtils.asObject(spec, SecretRequestWrapper.class);
-
     validateRequestPayload(dto);
+
+    if (!Objects.equals(orgIdentifier, dto.getSecret().getOrgIdentifier())
+        || !Objects.equals(projectIdentifier, dto.getSecret().getProjectIdentifier())) {
+      throw new InvalidRequestException("Invalid request, scope in payload and params do not match.", USER);
+    }
+
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountIdentifier, orgIdentifier, projectIdentifier), Resource.NONE, SECRET_EDIT_PERMISSION);
 
     return ResponseDTO.newResponse(ngSecretService.createFile(accountIdentifier, dto.getSecret(), uploadedInputStream));
   }
@@ -224,7 +286,22 @@ public class NGSecretResourceV2 {
   @ApiOperation(hidden = true, value = "Get Encryption Details", nickname = "postEncryptionDetails")
   @InternalApi
   public ResponseDTO<List<EncryptedDataDetail>> getEncryptionDetails(
-      NGAccessWithEncryptionConsumer ngAccessWithEncryptionConsumer) {
+      @NotNull NGAccessWithEncryptionConsumer ngAccessWithEncryptionConsumer) {
+    NGAccess ngAccess = ngAccessWithEncryptionConsumer.getNgAccess();
+    DecryptableEntity decryptableEntity = ngAccessWithEncryptionConsumer.getDecryptableEntity();
+    if (ngAccess == null || decryptableEntity == null) {
+      return ResponseDTO.newResponse(new ArrayList<>());
+    }
+    for (Field field : decryptableEntity.getSecretReferenceFields()) {
+      try {
+        SecretRefData secretRefData = (SecretRefData) field.get(decryptableEntity);
+        accessControlClient.checkForAccessOrThrow(ResourceScope.of(ngAccess.getAccountIdentifier(),
+                                                      ngAccess.getOrgIdentifier(), ngAccess.getProjectIdentifier()),
+            Resource.of(SECRET_RESOURCE_TYPE, secretRefData.getIdentifier()), SECRET_ACCESS_PERMISSION);
+      } catch (IllegalAccessException illegalAccessException) {
+        log.error("Error while checking access permission for secret: {}", field, illegalAccessException);
+      }
+    }
     return ResponseDTO.newResponse(secretManagerClientService.getEncryptionDetails(
         ngAccessWithEncryptionConsumer.getNgAccess(), ngAccessWithEncryptionConsumer.getDecryptableEntity()));
   }
